Base Update - Removal of extra Profile fields
I'm 95% sure this works. Lots of commenting that needs to be removed later, but could be useful if I decide to use the profiles for extra information but for now it was getting in the way of other plugins like BuddyPress which many people do use. Did actually test this on the Fabius install.